pairwithme/requirements.txt
Django==1.11.2
django-filter==1.0.4
django-webpack-loader==0.5.0
djangorestframework==3.6.3
mysqlclient==1.3.6
Pygments==2.2.0
pytz==2017.2
